feat(server): 新增雪花算法ID生成逻辑

+++ floruit-server/src/main/java/org/laz/floruitid/floruitserver/registrycenter/AbstractRegistryCenter.java
@@ -14,7 +14,7 @@ public abstract class AbstractRegistryCenter implements RegistryCenter {
     public abstract void initRegistryCenter();
 
     @Override
-    public abstract int getWorkerId();
+    public abstract long getWorkerId();
 
     @Override
     public abstract void cacheWorkerId();
